Converted AQI value from int to double.

+++ air-data-service/src/main/java/com/cobanoglu/airdataservice/model/AirQualityResponse.java
@@ -36,6 +36,6 @@ public class AirQualityResponse {
     @Data
     public static class MainData {
         @NotNull
-        private Integer aqi;
+        private Double aqi;
     }
 }
